requirements.txt
# Core
django==4.2.9
python-dotenv==1.0.1
str2bool==1.1

# UI
django-admin-adminlte==1.0.5

# Deployment
whitenoise==6.6.0
gunicorn==21.2.0

psycopg2-binary
# mysqlclient

requests

django-bootstrap5

celery[redis]

django_celery_beat

django-cors-headers

django_redis

python-binance

djangorestframework

pybit

channels

channels-redis

ccxt

daphne

djangorestframework-simplejwt